// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  Customer: 'Customer',
  Provider: 'Provider',
  Admin: 'Admin'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const Status = {
  activate: 'activate',
  suspend: 'suspend'
} as const

export type Status = (typeof Status)[keyof typeof Status]


export const PaymentStatus = {
  PAID: 'PAID',
  UNPAID: 'UNPAID',
  FREE: 'FREE'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const DietaryPreference = {
  HALAL: 'HALAL',
  VEGAN: 'VEGAN',
  VEGETARIAN: 'VEGETARIAN',
  ANY: 'ANY',
  GLUTEN_FREE: 'GLUTEN_FREE',
  KETO: 'KETO',
  PALEO: 'PALEO',
  DAIRY_FREE: 'DAIRY_FREE',
  NUT_FREE: 'NUT_FREE',
  LOW_SUGAR: 'LOW_SUGAR'
} as const

export type DietaryPreference = (typeof DietaryPreference)[keyof typeof DietaryPreference]


export const Cuisine = {
  BANGLEDESHI: 'BANGLEDESHI',
  ITALIAN: 'ITALIAN',
  CHINESE: 'CHINESE',
  INDIAN: 'INDIAN',
  MEXICAN: 'MEXICAN',
  THAI: 'THAI',
  JAPANESE: 'JAPANESE',
  FRENCH: 'FRENCH',
  MEDITERRANEAN: 'MEDITERRANEAN',
  AMERICAN: 'AMERICAN',
  MIDDLE_EASTERN: 'MIDDLE_EASTERN'
} as const

export type Cuisine = (typeof Cuisine)[keyof typeof Cuisine]


export const MealsStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type MealsStatus = (typeof MealsStatus)[keyof typeof MealsStatus]


export const OrderStatus = {
  PLACED: 'PLACED',
  PREPARING: 'PREPARING',
  READY: 'READY',
  DELIVERED: 'DELIVERED',
  CANCELLED: 'CANCELLED'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]


export const ReviewStatus = {
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type ReviewStatus = (typeof ReviewStatus)[keyof typeof ReviewStatus]